Refactor user endpoint

+++ src/main/kotlin/com/example/restservice/UserRepository.kt
@@ -2,11 +2,4 @@ package com.example.restservice
 
 import org.springframework.data.repository.CrudRepository
 
-
-// This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
-// CRUD refers Create, Read, Update, Delete
-
-// This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
-// CRUD refers Create, Read, Update, Delete
-interface UserRepository : CrudRepository<User?, Int?>
-
+interface UserRepository : CrudRepository<User?, Long?>
